\name{CNOlistPB}
\docType{data}
\alias{CNOlistPB}
\title{Toy data}
\description{This data object contains the data associated with the toy model example from the package vignette, already loaded and formatted as a CNOlist object.}
\usage{CNOlistPB}
\format{CNOlistPB is a list with the fields "namesCues" (character vector), "namesStimuli" (character vector), "namesInhibitors" (character vector), "namesSignals" (character vector), "timeSignals" (numerical vector), "valueCues" (numerical matrix), "valueInhibitors" (numerical matrix), "valueStimuli"(numerical matrix), "valueSignals"(numerical matrix).}
\source{
This data was generated with the CellNOptR add-on package CNORode. Full details of the data can be found in the reference below.
}
\references{
A. MacNamara, C. Terfve, D. Henriques, B. Penalver Bernabe and J. Saez-Rodriguez, 2012. State-time spectrum of signal transduction logic models. Physical biology, 9(4), p.045003.
}
\keyword{datasets}