\name{Model}
\docType{data}
\alias{ToyModelPB}
\title{Model}
\description{This data object contains the toy model from the package vignette, already loaded and formatted as a Model object.}
\usage{Model}
\format{ToyModel is a list with fields "reacID" (character vector), "namesSpecies" (character vector), "interMat" (numerical matrix), and "notMat"(numerical matrix). }
\source{This data and model is from the Physical Biology tutorial, "State-time spectrum of signal transduction logic models". It is used to demonstrate the assumptions and limitations of different logic model formalisms.}
\references{
A. MacNamara, C. Terfve, D. Henriques, B. Penalver Bernabe and J. Saez-Rodriguez. State-time spectrum of signal transduction logic models. Physical Biology, 2012. Submitted.
}
\keyword{datasets}
